refactor: 상품 목록 관련 DTO 구조 개선을 위한 패키지 분리 및 참조 수정
