+++ app/src/main/java/woowacourse/shopping/backend/mock/InAppMockShoppingServer.kt
@@ -9,12 +9,12 @@ import okhttp3.mockwebserver.MockWebServer
 import okhttp3.mockwebserver.RecordedRequest
 import woowacourse.shopping.backend.retrofit.dto.CartQuantity
 import woowacourse.shopping.backend.retrofit.dto.CartRequest
-import woowacourse.shopping.backend.retrofit.dto.Content
+import woowacourse.shopping.backend.retrofit.dto.productlist.Content
 import woowacourse.shopping.backend.retrofit.dto.OrderInfo
 import woowacourse.shopping.backend.retrofit.dto.Pageable
 import woowacourse.shopping.backend.retrofit.dto.PageableResponse
-import woowacourse.shopping.backend.retrofit.dto.Product
-import woowacourse.shopping.backend.retrofit.dto.ProductResponse
+import woowacourse.shopping.backend.retrofit.dto.productlist.Product
+import woowacourse.shopping.backend.retrofit.dto.productlist.ProductResponse
 import woowacourse.shopping.backend.retrofit.dto.ShoppingCartPageable
 import woowacourse.shopping.backend.retrofit.dto.ShoppingCartResponse
 import woowacourse.shopping.backend.retrofit.dto.Sort
