+++ app/src/main/java/woowacourse/shopping/backend/retrofit/dto/productlist/Content.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.backend.retrofit.dto
+package woowacourse.shopping.backend.retrofit.dto.productlist
 
 import kotlinx.serialization.Serializable
